+++ CHANGELOG.rst
@@ -11,7 +11,7 @@ Changelog
 - (Changed) ``Interpreter.__init__`` does not anymore stabilize the statechart. Stabilization is done during the
   first call of ``execute_once``.
 - (Changed) ``Story.tell`` returns a list of ``MacroStep`` (the *trace*) instead of an ``Interpreter`` instance.
-- (Changed) The name of some attributes of an event in a tester story changes (e.g. *event* becomes *consumed event*,
+- (Changed) The name of some attributes of an event in a tester story changes (e.g. *event* becomes *consumed_event*,
   *state* becomes *entered_state* or *exited_state* or *source_state* or *target_state*).
 - (Removed) ``Interpreter.trace``, as it can be easily obtained from ``execute_once`` or using ``log``.
 - (Removed) ``Interpreter.__init__`` does not accept an ``initial_time`` parameter.